Rename validate command to check
